% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-LandsepiParams.R
\name{checkCultivars}
\alias{checkCultivars}
\title{Check cultivars}
\usage{
checkCultivars(params)
}
\arguments{
\item{params}{a LandsepiParams object.}
}
\value{
a boolean, TRUE if OK, FALSE otherwise
}
\description{
check cultivars validity
}
