% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-LandsepiParams.R
\name{checkCultivarsGenes}
\alias{checkCultivarsGenes}
\title{Check cultivars genes}
\usage{
checkCultivarsGenes(params)
}
\arguments{
\item{params}{a LandsepiParams object.}
}
\value{
a boolean, TRUE if OK, FALSE otherwise
}
\description{
Checks CultivarsGene data.frame validity
}
