% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-LandsepiParams.R
\name{checkDispersalHost}
\alias{checkDispersalHost}
\title{Check host dispersal}
\usage{
checkDispersalHost(params)
}
\arguments{
\item{params}{a LandsepiParams Object.}
}
\value{
a boolean TRUE if OK, FALSE otherwise
}
\description{
Checks host dispersal matrix validity.
}
