% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-LandsepiParams.R
\name{checkDispersalPathogen}
\alias{checkDispersalPathogen}
\title{Check pathogen dispersal}
\usage{
checkDispersalPathogen(params)
}
\arguments{
\item{params}{a LandsepiParams Object.}
}
\value{
a boolean TRUE if OK, FALSE otherwise
}
\description{
Checks pathogen dispersal validity
}
