% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-LandsepiParams.R
\name{checkGenes}
\alias{checkGenes}
\title{Check genes}
\usage{
checkGenes(params)
}
\arguments{
\item{params}{a LandsepiParams object.}
}
\value{
a boolean, TRUE if OK, FALSE otherwise
}
\description{
checks Genes data.frame validity
}
