% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-LandsepiParams.R
\name{checkInoculum}
\alias{checkInoculum}
\title{Check inoculum}
\usage{
checkInoculum(params)
}
\arguments{
\item{params}{a LandsepiParams object.}
}
\value{
a boolean, TRUE if OK, FALSE otherwise
}
\description{
Checks inoculum validity.
}
