% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-LandsepiParams.R
\name{checkLandscape}
\alias{checkLandscape}
\title{Check the landscape}
\usage{
checkLandscape(params)
}
\arguments{
\item{params}{a LandsepiParams Object.}
}
\value{
TRUE if Ok, FALSE otherwise
}
\description{
Checks landscape validity
}
