% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-LandsepiParams.R
\name{checkOutputs}
\alias{checkOutputs}
\title{Check outputs}
\usage{
checkOutputs(params)
}
\arguments{
\item{params}{a LandsepiParams object.}
}
\value{
a boolean, TRUE if OK, FALSE otherwise
}
\description{
Checks outputs validity.
}
