% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-LandsepiParams.R
\name{checkPathogen}
\alias{checkPathogen}
\title{Check pathogen}
\usage{
checkPathogen(params)
}
\arguments{
\item{params}{a LandsepiParams Object.}
}
\value{
a boolean, TRUE if OK, FALSE otherwise
}
\description{
Checks papthogen validity
}
