% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-LandsepiParams.R
\name{checkSimulParams}
\alias{checkSimulParams}
\title{Check simulation parameters}
\usage{
checkSimulParams(params)
}
\arguments{
\item{params}{a LandsepiParams Object.}
}
\value{
TRUE if OK for simulation, FALSE otherwise
}
\description{
Checks validity of a LandsepiParams object.
}
