% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-LandsepiParams.R
\name{checkTime}
\alias{checkTime}
\title{Check time}
\usage{
checkTime(params)
}
\arguments{
\item{params}{a LandsepiParams Object.}
}
\value{
a boolean TRUE if times are setted.
}
\description{
Checks time parameters validity
}
