% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-LandsepiParams.R
\name{checkTreatment}
\alias{checkTreatment}
\title{Check treatment}
\usage{
checkTreatment(params)
}
\arguments{
\item{params}{a LandsepiParams Object.}
}
\value{
a boolean, TRUE if OK, FALSE otherwise
}
\description{
Checks treatment validity
}
