% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPKGTools.R
\name{createLandscapeGPKG}
\alias{createLandscapeGPKG}
\title{createLandscapeGPKG}
\usage{
createLandscapeGPKG(landscape, outputfile)
}
\arguments{
\item{landscape}{a sf object}

\item{outputfile}{GPKG output file name}
}
\value{
the outputfile name (full path)
}
\description{
Creates a GPKG file from a sf object
}
\details{
Generates a GPKG file with two layouts.
One layout called "croptypeID" containing the croptypes ID for every year.
One another layout called "area" containing the area of every polygon.
}
