% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-LandsepiParams.R
\name{CroptypeBDD2Params}
\alias{CroptypeBDD2Params}
\title{CroptypeBDD2Params}
\usage{
CroptypeBDD2Params(inputGPKG)
}
\arguments{
\item{inputGPKG}{a GPKG filename}
}
\value{
a data.frame LandsepiParams@Croptypes compatible
}
\description{
Converts BDD table to Croptype LandspeParams object
}
