% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-LandsepiParams.R
\name{CultivarBDD2Params}
\alias{CultivarBDD2Params}
\title{CultivarBDD2Params}
\usage{
CultivarBDD2Params(inputGPKG)
}
\arguments{
\item{inputGPKG}{a GPKG filename}
}
\value{
a data.frame LandsepiParams@Cultivars compatible
}
\description{
Converts BDD table Cultivar to a Cultivar LandsepiParams object
}
