% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-LandsepiParams.R
\name{CultivarGeneBDD2Params}
\alias{CultivarGeneBDD2Params}
\title{CultivarGeneBDD2Params}
\usage{
CultivarGeneBDD2Params(inputGPKG)
}
\arguments{
\item{inputGPKG}{a GPKG filename}
}
\value{
a data.frame LandsepiParams@CultivarsGenes compatible
}
\description{
Converts BDD table to LandsepiParams object
}
