% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cultivars_List.R
\docType{data}
\name{Cultivars_list}
\alias{Cultivars_list}
\title{Cultivars Type list}
\format{
A list of list indexed by type name
\itemize{
\item cultivarName: cultivar names (cannot accept space),
\item initial_density: host densities (per square meter) at the beginning of the cropping season
as if cultivated in pure crop,
\item max_density: maximum host densities (per square meter) at the end of the cropping season
as if cultivated in pure crop,
\item growth rate: host growth rates,
\item reproduction rate: host reproduction rates,
\item yield_H: theoretical yield (in weight or volume units / ha / cropping season)
associated with hosts in sanitary status H as if cultivated in pure crop,
\item yield_L: theoretical yield (in weight or volume units / ha / cropping season)
associated with hosts in sanitary status L as if cultivated in pure crop,
\item yield_I: theoretical yield (in weight or volume units / ha / cropping season)
associated with hosts in sanitary status I as if cultivated in pure crop,
\item yield_R: theoretical yield (in weight or volume units / ha / cropping season)
associated with hosts in sanitary status R as if cultivated in pure crop,
\item planting_cost = planting costs (in monetary units / ha / cropping season) as if cultivated in pure crop,
\item market_value = market values of the production (in monetary units / weight or volume unit).
}
}
\usage{
Cultivars_list
}
\description{
A set of configurated cultivars types
}
\keyword{datasets}
