% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_landsepi.R
\name{demo_landsepi}
\alias{demo_landsepi}
\title{Package demonstration}
\usage{
demo_landsepi(seed = 12345)
}
\arguments{
\item{seed}{an integer used as seed value (for random number generator)}
}
\description{
Run a demonstration of the package.
}
\details{
Run demo landsepi

Run a simulated example of mosaic deployment strategy of two resistant cultivars in balanced proportions and 
high level of spatial aggregation. The generated model outputs are graphics and text files.
}
