\name{dispH}
\alias{dispH}
\docType{data}
\title{
  Host dispersal data for testing.
}
\description{
  A vectorised dispersal matrix of the host.
}
\usage{dispH}
\format{
  The format is:
 num [1:24025] 1 0 0 0 0 0 0 0 0 0 ...
}
\details{
  In this example, the host is a cultivated plant: seeds are harvested and do not disperse. Thus the dispersal matrix is the identity matrix.
}
\examples{
dispH
summary(dispH)
}
\keyword{datasets}
