% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-LandsepiParams.R
\name{GeneBDD2Params}
\alias{GeneBDD2Params}
\title{GeneBDD2Params}
\usage{
GeneBDD2Params(inputGPKG)
}
\arguments{
\item{inputGPKG}{a LandsepiParams}
}
\value{
a data.frame LandsepiParams@Genes compatible
}
\description{
Converts BDD table Gene to LandsepiParams object
}
