% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPKGTools.R
\name{getGPKGArea}
\alias{getGPKGArea}
\title{getGPKGArea}
\usage{
getGPKGArea(gpkgfile)
}
\arguments{
\item{gpkgfile}{a GPKG file}
}
\value{
a vector of the area of each polygons
}
\description{
Gets "area" layer as a vector of a GPKG file
}
