% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPKGTools.R
\name{getGPKGCroptypes}
\alias{getGPKGCroptypes}
\title{getGPKGCroptypes}
\usage{
getGPKGCroptypes(inputGPKG)
}
\arguments{
\item{inputGPKG}{a GPKG filename}
}
\value{
a data.frame with croptypeID, CultivarNames, and proportions
}
\description{
Gets Croptypes and Cultivars proportions from a GPKG file
}
