% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPKGTools.R
\name{getGPKGCroptypesRaw}
\alias{getGPKGCroptypesRaw}
\title{getGPKGCroptypesRaw}
\usage{
getGPKGCroptypesRaw(inputGPKG)
}
\arguments{
\item{inputGPKG}{a GPKG filename}
}
\value{
a data.frame with croptypeID, CultivarID, and proportions
}
\description{
Gets Croptypes and Cultivars proportions from a GPKG file without refactoring
}
