% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPKGTools.R
\name{getGPKGCultivars}
\alias{getGPKGCultivars}
\title{getGPKGCultivars}
\usage{
getGPKGCultivars(inputGPKG)
}
\arguments{
\item{inputGPKG}{a GPKG filename}
}
\value{
a data.frame
}
\description{
Gets Cultivars from a GPKG file
}
