% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPKGTools.R
\name{getGPKGCultivarsGenes}
\alias{getGPKGCultivarsGenes}
\title{getGPKGCultivarsGenes}
\usage{
getGPKGCultivarsGenes(inputGPKG)
}
\arguments{
\item{inputGPKG}{a GPKG filename}
}
\value{
a data.frame
}
\description{
Gets Cultivars Genes from a GPKG file
}
