% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPKGTools.R
\name{getGPKGGeneIDForCultivar}
\alias{getGPKGGeneIDForCultivar}
\title{getGPKGGeneIDForCultivar}
\usage{
getGPKGGeneIDForCultivar(inputGPKG, cultivarID)
}
\arguments{
\item{inputGPKG}{a GPKG filename}

\item{cultivarID}{a cultivarID}
}
\value{
a data.frame of GeneID
}
\description{
Gets GeneID of a cultivar
}
