% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPKGTools.R
\name{getGPKGGenes}
\alias{getGPKGGenes}
\title{getGPKGGenes}
\usage{
getGPKGGenes(inputGPKG)
}
\arguments{
\item{inputGPKG}{a GPKG filename}
}
\value{
a data.frame
}
\description{
Gets Genes from a GPKG file
}
