% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPKGTools.R
\name{getGPKGRotation}
\alias{getGPKGRotation}
\title{getGPKGRotation}
\usage{
getGPKGRotation(gpkgfile)
}
\arguments{
\item{gpkgfile}{a GPKG file}
}
\value{
a matrix as rows for polygons and cols for years
}
\description{
Gets Croptypes ID rotation years as a matrix
}
