% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPKGTools.R
\name{GPKGAddInputData}
\alias{GPKGAddInputData}
\title{GPKGAddInputData}
\usage{
GPKGAddInputData(
  gpkgfile,
  table = "",
  data = data.frame(),
  deleteExistingData = FALSE
)
}
\arguments{
\item{gpkgfile}{a gpkg filename}

\item{table}{table name}

\item{data}{a data to write in BDD, should be the return of a function param2XXXXXBDD}

\item{deleteExistingData}{if TRUE overwrite data if already present in gpkg file, default FALSE}
}
\description{
GPKGAddInputData
}
\details{
Adds data 'data' values in the table 'table' using 'data' colnames
}
