% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPKGTools.R
\name{GPKGAddTables}
\alias{GPKGAddTables}
\title{GPKGAddTables}
\usage{
GPKGAddTables(gpkgfile)
}
\arguments{
\item{gpkgfile}{a GPKF file}
}
\value{
the GPKG file name
}
\description{
Adds non spatial data table definitions (sqlite) into a GPKG file.
It adds Cultivar, CultivarList, Gene, GeneList tables
}
