% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Math-Functions.R
\name{invlogit}
\alias{invlogit}
\title{Inverse logit function}
\usage{
invlogit(x)
}
\arguments{
\item{x}{a numeric object}
}
\value{
An object of the same type as x containing the invlogits of the input values.
}
\description{
Given a numeric object, return the invlogit of the values. Missing values (NAs) are allowed.
}
\details{
The invlogit is defined by exp(x) / (1+exp(x)). Values in x of -Inf or Inf return invlogits
of 0 or 1 respectively.
Any NAs in the input will also be NAs in the output.
}
\examples{
invlogit(10)
}
