% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{is.in.01}
\alias{is.in.01}
\title{is.in.01}
\usage{
is.in.01(x, exclude0 = FALSE)
}
\arguments{
\item{x}{a number or vector or matrix}

\item{exclude0}{TRUE is 0 is excluded, FALSE otherwise (default)}
}
\value{
a logical of the same size as x
}
\description{
Tests if a number or vector is in the interval [0,1]
}
\examples{
is.in.01(-5)
is.in.01(0)
is.in.01(1)
is.in.01(0, exclude0 = TRUE)
is.in.01(2.5)
is.in.01(matrix(5:13/10, nrow=3))
}
