% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{is.positive}
\alias{is.positive}
\title{is.positive}
\usage{
is.positive(x)
}
\arguments{
\item{x}{a number or vector or matrix}
}
\value{
a logical of the same size as x
}
\description{
Tests if a number or vector is positive (including 0)
}
\examples{
is.positive(-5)
is.positive(10)
is.positive(2.5)
is.positive(matrix(1:9, nrow=3))
}
