% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{is.strict.positive}
\alias{is.strict.positive}
\title{is.strict.positive}
\usage{
is.strict.positive(x)
}
\arguments{
\item{x}{a number or vector or matrix}
}
\value{
a logical of the same size as x
}
\description{
Tests if a number or vector is strictly positive (i.e. excluding 0)
}
\examples{
is.strict.positive(-5)
is.strict.positive(10)
is.strict.positive(2.5)
is.strict.positive(matrix(1:9, nrow=3))
}
