% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{is.wholenumber}
\alias{is.wholenumber}
\title{is.wholenumber}
\usage{
is.wholenumber(x, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{a number or vector or matrix}

\item{tol}{double tolerance}
}
\value{
a logical of the same format as x
}
\description{
Tests if a number or vector is a whole number
}
\examples{
is.wholenumber(-5)
is.wholenumber(10)
is.wholenumber(2.5)
is.wholenumber(matrix(1:9, nrow=3))
}
