% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-LandsepiParams.R
\name{loadLandscape}
\alias{loadLandscape}
\title{Load a landscape}
\usage{
loadLandscape(id = 1)
}
\arguments{
\item{id}{a landscape ID between 1 to 5 (default = 1)}
}
\value{
a landscape in sp format
}
\description{
Loads one of the five built-in landscapes simulated using a T-tesselation algorithm
and composed of 155, 154, 152, 153 and 156 fields, respectively.
Each landscape is identified by a numeric from 1 to 5.
}
\examples{
land <- loadLandscape(1)
length(land)
}
\seealso{
\link{landscapeTEST}, \link{setLandscape}
}
