% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-LandsepiParams.R
\name{loadSimulParams}
\alias{loadSimulParams}
\title{Load simulation parameters}
\usage{
loadSimulParams(inputGPKG = "")
}
\arguments{
\item{inputGPKG}{name of the GPKG file.}
}
\value{
a LandsepiParams object.
}
\description{
Loads a GPKG file from the output of a landsepi simulation.
}
\details{
See \code{\link{saveDeploymentStrategy}}.
}
