% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-LandsepiParams.R
\name{loadTreatment}
\alias{loadTreatment}
\title{Load treatment parameters}
\usage{
loadTreatment(disease = "no pathogen")
}
\arguments{
\item{disease}{a disease name, among "mildew", "sigatoka" and "no pathogen"}
}
\value{
a list of treatment parameters:
\itemize{
\item treatment_degradation_rate = degradation rate (per time step) of chemical concentration,
\item treatment_efficiency = maximal efficiency of chemical treatments (i.e. fractional reduction
of pathogen infection rate at the time of application),
\item treatment_timesteps = vector of time steps corresponding to treatment application dates,
\item treatment_cultivars = vector of indices of the cultivars that receive treatments,
\item treatment_cost = cost of a single treatment application (monetary units/ha)
\item treatment_application_threshold = vector of thresholds (i.e. disease severity, one
for each treated cultivar) above which the treatment is applied in a polygon.
}
}
\description{
Loads a list of treatment parameters for a specific disease (initialised at 0
, i.e. absence of treatments)
}
\details{
Chemical treatment is applied in a polygon only if disease severity (i.e. I/N) in
this polygon exceeds the threshold given by \code{treatment_application_threshold}.
Treatment efficiency is maximum (i.e. equal to the parameter treatment_efficiency)
at the time of treatment application (noted \eqn{t*}); then it decreases with time
(i.e. natural pesticide degradation) and host growth (i.e. new biomass is not protected
by treatments):                                                                                                                                                                                     protected by treatments):Efficiency of the treatment at time t after the application date is given by:
\eqn{ efficiency(t) = treatment\_efficiency / (1 + exp(a-b*C(t))) }
with \eqn{ C(t)= C_1 * C_2}: \itemize{
\item{\eqn{C_1 = exp(- treatment\_degradation\_rate * \Delta t) } is the reduction of
fungicide concentration due to time (e.g. natural degradation, volatilization, weathering),
with \eqn{\Delta t = t - t*} the timelag passed since the time of
treatment application.}
\item{ \eqn{ C_2 = min(1, N(t*) / N(t)) } is the reduction of fungicide concentration due
to plant growth, since new plant tissue is not covered by fungicide.
\eqn{N(t*)} and \eqn{N(t)} being the number of
host individuals  a the time of treatment \eqn{t*} and at time \eqn{t}, respectively.}
\item{ \eqn{a \in [3.5 ; 4.5]} and \eqn{b \in [8 ; 9]} are shape parameters.}
}
}
\examples{
treat <- loadTreatment("sigatoka")
treat
}
\seealso{
\link{setTreatment}
}
