% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Math-Functions.R
\name{logit}
\alias{logit}
\title{Logit function}
\usage{
logit(x)
}
\arguments{
\item{x}{a numeric object containing values between 0 and 1}
}
\value{
An object of the same type as x containing the logits of the input values.
}
\description{
Given a numeric object, return the logit of the values. Missing values (NAs) are allowed.
}
\details{
The logit is defined by log(x/(1-x)). Values in x of 0 or 1 return logits of -Inf or Inf respectively.
Any NAs in the input will also be NAs in the output.
}
\examples{
logit(0.5)
}
