% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-LandsepiParams.R
\name{params2CroptypeBDD}
\alias{params2CroptypeBDD}
\title{params2CroptypeBDD}
\usage{
params2CroptypeBDD(params)
}
\arguments{
\item{params}{a LandsepiParams object.}
}
\value{
a data.frame BDD compatible
}
\description{
Converts a LandsepiParams object to a value compatible with BDD croptype Table
}
