% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-LandsepiParams.R
\name{params2CultivarBDD}
\alias{params2CultivarBDD}
\title{params2CultivarBDD}
\usage{
params2CultivarBDD(params)
}
\arguments{
\item{params}{a LandsepiParam object.}
}
\value{
a data.frame BDD compatible
}
\description{
Converts a LandsepiParams object to a value compatible with BDD cultivar Table
}
