% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-LandsepiParams.R
\name{params2GeneBDD}
\alias{params2GeneBDD}
\title{params2GeneBDD}
\usage{
params2GeneBDD(params)
}
\arguments{
\item{params}{a LandsepiParam object.}
}
\value{
a data.frame BDD compatible
}
\description{
Converts a LandsepiParams object to a value compatible with BDD Gene Table
}
