% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-LandsepiParams.R
\name{params2GeneListBDD}
\alias{params2GeneListBDD}
\title{params2GeneListBDD}
\usage{
params2GeneListBDD(params)
}
\arguments{
\item{params}{a LandsepiParam object.}
}
\value{
a data.frame BDD compatible
}
\description{
Converts a LandsepiParams object to a value compatible with BDD cultivarsGenes Table
}
