% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-LandsepiParams.R
\name{print}
\alias{print}
\alias{print,LandsepiParams-method}
\title{print}
\usage{
\S4method{print}{LandsepiParams}(x, ...)
}
\arguments{
\item{x}{a LandsepiParams object}

\item{...}{print options}
}
\description{
Prints a LandespiParams object.
}
