% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-LandsepiParams.R
\name{resetCultivarsGenes}
\alias{resetCultivarsGenes}
\title{Reset cultivars genes}
\usage{
resetCultivarsGenes(params)
}
\arguments{
\item{params}{a LandsepiParams object.}
}
\value{
a LandsepiParams object
}
\description{
Resets the lists of genes carried by all cultivars
}
