% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShiny.R
\name{runShinyApp}
\alias{runShinyApp}
\title{runShinyApp}
\usage{
runShinyApp()
}
\description{
Launches landsepi shiny application into browser
}
\details{
R packages needed to run the shiny app :
install.packages(c("shiny","DT", "shinyjs", "gridExtra", "png", "grid", "future", "promises", "tools"))
}
