% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-LandsepiParams.R
\name{setDispersalHost}
\alias{setDispersalHost}
\title{Set host dispersal}
\usage{
setDispersalHost(params, mat)
}
\arguments{
\item{params}{a LandsepiParams Object.}

\item{mat}{a square matrix giving the probability of host dispersal
from any field of the landscape to any other field.
It can be generated manually, or, alternatively, via \code{\link{loadDispersalHost}}.
The size of the matrix must match the number of fields in the landscape.}
}
\value{
a LandsepiParam object.
}
\description{
Updates a LandsepiParams object with a host dispersal matrix.
}
\details{
the dispersal matrix gives the probability for a host individual in a field i (row)
to migrate to field j (column) through dispersal.
If the host is a cultivated plant: seeds are harvested and do not disperse.
Thus the dispersal matrix is the identity matrix.
}
\examples{
\dontrun{
simul_params <- createSimulParams()
simul_params <- setLandscape(simul_params, loadLandscape(1))
d <- loadDispersalHost(simul_params)
simul_params <- setDispersalHost(simul_params, d)
simul_params@DispHost
}
}
\seealso{
\link{loadDispersalHost}
}
