% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-LandsepiParams.R
\name{setReproSexProb}
\alias{setReproSexProb}
\title{set a vector of probabilities for an infectious host to reproduce via sex rather
than via cloning.}
\usage{
setReproSexProb(params, vec)
}
\arguments{
\item{params}{a LandsepiParams object}

\item{vec}{a vector of size TimeParam$nTSpY +1 (season end) with the probabilities for an infectious host
to reproduce via sex rather than via cloning at each step.}
}
\value{
a LandsepiParams object updated
}
\description{
set the probabilities for an infectious host to reproduce via sex rather
than via cloning at every timestep (\code{TimeParam$nTSpY}).
}
