% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-LandsepiParams.R
\name{setSeed}
\alias{setSeed}
\title{Set the seed}
\usage{
setSeed(params, seed)
}
\arguments{
\item{params}{a LandsepiParams Object.}

\item{seed}{an integer used as seed value (for random number generator).}
}
\value{
a LandsepiParams object.
}
\description{
Updates a LandsepiParams object with a seed value for random number generator
}
\examples{
\dontrun{
simul_params <- createSimulParams()
simul_params <- setSeed(simul_params, 100)
simul_params@Seed
}
}
