% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{setSeedValue}
\alias{setSeedValue}
\title{setSeedValue}
\usage{
setSeedValue(seed = NULL)
}
\arguments{
\item{seed}{an interger as seed value or NULL}
}
\value{
the new seed value for RNG
}
\description{
Set RNG seed to seed value if not NULL, otherwise set it to timestamps value
}
\details{
Sets seed for "Mersenne-Twister" algorithm using Inversion generation
}
\examples{
setSeedValue(seed = 10)
}
