% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-LandsepiParams.R
\name{show}
\alias{show}
\alias{show,LandsepiParams-method}
\title{show}
\usage{
\S4method{show}{LandsepiParams}(object)
}
\arguments{
\item{object}{a LandsepiParams object}
}
\description{
Shows a LandespiParams object.
}
