% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-LandsepiParams.R
\name{summary}
\alias{summary}
\alias{summary,LandsepiParams-method}
\title{summary}
\usage{
\S4method{summary}{LandsepiParams}(object)
}
\arguments{
\item{object}{a LandsepiParams object.}
}
\description{
Prints the summary of a LandespiParams object.
}
