% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{switch_patho_to_aggr}
\alias{switch_patho_to_aggr}
\title{Switch from index of genotype to indices of agressiveness on different components}
\usage{
switch_patho_to_aggr(index_patho, Ngenes, Nlevels_aggressiveness)
}
\arguments{
\item{index_patho}{index of pathogen genotype}

\item{Ngenes}{number of resistance genes}

\item{Nlevels_aggressiveness}{vector of the number of adaptation levels related to each resistance gene}
}
\value{
a vector containing the indices of aggressiveness on the different components targeter by the resistance genes
}
\description{
Finds the level of aggressiveness on different components (targeted by different resistance genes)
from the index of a given pathogen genotype
}
\examples{
switch_patho_to_aggr(5, 3, c(2, 2, 3))
}
