% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-LandsepiParams.R
\name{loadPathogen}
\alias{loadPathogen}
\title{Load pathogen parameters}
\usage{
loadPathogen(disease = "rust")
}
\arguments{
\item{disease}{a disease name (default: "rust")}
}
\value{
a list of pathogen aggressiveness parameters on a susceptible host
for a pathogen genotype not adapted to resistance
}
\description{
Loads default pathogen parameters for a specific disease
}
\details{
Available diseases:
\itemize{
\item "rust"
}
}
\examples{
basic_patho_params <- loadPathogen()
basic_patho_params
}
\seealso{
\link{setPathogen}
}
