% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-LandsepiParams.R
\name{loadTreatment}
\alias{loadTreatment}
\title{Load treatment parameters}
\usage{
loadTreatment()
}
\value{
a list of treatment parameters:
\itemize{
\item treatment_reduction_rate = reduction per time step of treatment concentration,
\item treatment_efficiency = maximal efficiency of chemical treatments (i.e. fractional reduction of pathogen infection rate at the application date),
\item treatment_timesteps = vector of time-steps corresponding to treatment application dates,
\item treatment_cultivars  = vector of cultivar indices that receive treatments,
\item treatment_cost = cost of a single treatment application (monetary units/ha).
}
}
\description{
Load the list of treatment parameters required by the model (initialised at 0
, i.e. absence of treatments)
}
\examples{
treat <- loadTreatment()
treat
}
\seealso{
\link{setTreatment}
}
